chinese_handwriting_server/requirements.txt
certifi==2021.5.30
click==8.0.1
Flask==2.0.1
Flask-Cors==3.0.10
importlib-metadata==4.6.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.21.1
Pillow==8.3.1
six==1.16.0
torch==1.8.0+cu111
torchaudio==0.8.0
torchvision==0.9.0+cu111
typing-extensions==3.10.0.0
Werkzeug==2.0.1
zipp==3.5.0
